feat(metube): Update container to v2023-08 (#12413)
Updates the application container image to the latest version v2023-08.

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
It is NOT tested. I tried adding my repo fork in my TrueNAS installation
but this fails a long list of lines like
`templates.questions.versions.security.name: 'security' is not a valid
version name. [EINVAL]`. I also check the development docs and FAQ but
did not find any hints on how I can test this.

**📃 Notes:**

I'm unsure why this is not done automatically / by a bot but the image
was not updated in 6 months. I could not find any commits or issues that
would indicate that the old version is used on purpose.

This PR updates the image to the latest on build in [metube
truecharts-containers](https://github.com/truecharts/containers/blob/master/mirror/metube/Dockerfile)
and is uploaded to [metube
repository](https://quay.io/repository/truecharts/metube?tab=tags).

This PR does **NOT** fix the reason for why automated version update PRs
are not working. My assumption is that the bot does not understand the
image tags versioning schema and does not know that `v2023-08` is a
newer as `v2023-02`.

**✔️ Checklist:**

- [x] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`