Adding delay for startup probe in values.yaml of nginx_proxy_manager
If the current version of nginx_proxy_manager is deployed as TrueNAS Scale App the bootup fails due to the fail of the startup probe. Probably the startup time of either the maridb or the nginx proxy manager itself is too long leading to a failure of the startup probe script. Adding an initialDelaySeconds of 300 leads to a successful boot of the app.