feat(cluster-issuer): Add ability to create cluster-wide certificates (#13852)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->

This PR adds the next piece needed to support cluster-wide certificates
in Truecharts.
See PRs: 
- #13756 
- https://github.com/truecharts/library-charts/pull/537 
- https://github.com/truecharts/containers/pull/33273

Relevant issues:  #8634
This PR only adds what is needed in cluster-issuer. I will be pushing a
separate PR that will change the contents of the question includes.
⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [X] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

Tested it produces the correct yaml with helm:

```
helm template -n cluster-issuer cluster-issuer . 
```

Tested with 0, 1 and 2 certificate definitions in `values.yaml`. The
produced YAML is correct (with one exception - its missing the
`secretTemplate` property, which needs
https://github.com/truecharts/library-charts/pull/537 in order to work)

**📃 Notes:**
<!-- Please enter any other relevant information here -->
Should be merged after
https://github.com/truecharts/library-charts/pull/537 is merged and the
common version dependency is bumped to reflect the new common version.

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [X] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [X] ⚠️ My changes generate no new warnings
- [X] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: sdimovv <36302090+sdimovv@users.noreply.github.com>
Signed-off-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>
Signed-off-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Stavros Kois <47820033+stavros-k@users.noreply.github.com>
Co-authored-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>