Fixing metrics (#14719)
**Description**
Several *arr applications support metrics. However, they were missing a
`targetSelector` that points the `ServiceMonitor` resource to the
correct `service` which is named `metrics`.

It also fixes a bug for uptime-kuma where the metrics configuration
points to the port named `metrics` which does not exist. The metrics for
uptime-kuma are exposed via the regular service on the `main` port.

⚒️ Fixes  #14468

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [X] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Manually applied the changes to the `ServiceMonitor` resources using
`kubectl patch` to see if the metrics started working. They did, so
these changes will ensure the `ServiceMonitor` resources will be
deployed with the right `matchLabels` and `port` values.

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🪞 I have opened a PR on
[truecharts/containers](https://github.com/truecharts/containers) adding
the container to TrueCharts mirror repo.
- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._